Reorganized files.
Made separate files for class declaration and definition.
