+++ .gitignore
@@ -1,0 +1,4 @@
+2048.DSK
+CONSTANT.BAK
+DISPLAY.BAK
+GRID.BAK
